pyproject.toml
[project]
name = "qdrant-indexer"
version = "0.1.0"
description = "内部ドキュメントのセマンティック検索インデクサー CLI"
requires-python = ">=3.13"

dependencies = [
    "typer>=0.15.0",
    "fastembed>=0.6.0",
    "qdrant-client>=1.13.0",
    "pyyaml>=6.0.0",
]

# 開発専用の依存
[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "types-PyYAML>=6.0.0",
]

# CLI エントリポイント
[project.scripts]
qdrant-indexer = "qdrant_indexer.cli:app"

# ビルドシステム（src レイアウト）
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qdrant_indexer"]

# ── ruff ──────────────────────────────────────────────
[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",         # pyflakes
    "I",         # isort
    "N",         # pep8-naming
    "UP",        # pyupgrade
    "B",         # flake8-bugbear
    "C4",        # flake8-comprehensions
    "DTZ",       # flake8-datetimez
    "T10",       # flake8-debugger
    "ISC",       # flake8-implicit-str-concat
    "RET",       # flake8-return
    "SIM",       # flake8-simplify
    "TID",       # flake8-tidy-imports
    "PTH",       # flake8-use-pathlib
    "PL",        # pylint
]
ignore = ["E501", "PLR0913", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "N802", "PLC2401"]

[tool.ruff.lint.isort]
known-first-party = ["qdrant_indexer"]

# ── pytest ────────────────────────────────────────────
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# ── coverage ──────────────────────────────────────────
[tool.coverage.run]
source = ["qdrant_indexer"]
omit = ["*/tests/*", "*/conftest.py"]

[tool.coverage.report]
show_missing = true
precision = 2

# ── mypy ──────────────────────────────────────────────
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["fastembed.*", "qdrant_client.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
